refactor: decouple demos from searching_algorithms
